backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "procure-to-pay-backend"
version = "0.1.0"
description = "Procure-to-Pay Backend API"
dependencies = [
    "Django>=4.2,<5.0",
    "djangorestframework>=3.14.0",
    "djangorestframework-simplejwt>=5.3.0",
    "django-environ>=0.11.0",
    "psycopg2-binary>=2.9.0",
    "celery>=5.3.0",
    "redis>=5.0.0",
    "gunicorn>=21.2.0",
    "django-cors-headers>=4.3.0",
    "Pillow>=10.0.0",
    "dj-database-url>=2.1.0",
    "django-storages>=1.14.0",
    "boto3>=1.28.0", 
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-django>=4.5.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
]
